PLF-7757 : remove users from search index when there are not in /platform/users
All users profiles are indexed for search, except if they are disabled. So users not in /platform/users where also indexed whereas they are considered are disabled as well.
This fix do not index users when they are disabled or when they don't belong to the group /platform/users.
It updates the profile indexer and adds a membership listener to index/unindex an user profile when his permissions are changed.