SOC-6143 : set correct etag to avoid sending the same response again on /spaces
The etag on /spaces endpoint was not set correctly as is was calculated only on spaces id, not on all the /spaces response content.
Therefore a 304 http code was sent on subsequent /spaces calls if only properties of the same spaces changed.

This fix set the correct etag based on the whole content of the response.